Better command-line UI: default to export; or match <registry>
Run the code in one of two ways:

  lua extract-tlds.lua

will export the entire root db as a tab-delimited text file suitable for
importing into a Google Drive Sheet.

  lua extract-tlds.lua match <registry>

where <registry> is one of the entries in the registries table in the code, eg
"donuts"; this will print all the matching records for that registry.

It's easy to add more; I defined these because they were the big ones (single
companies with more than three or four domains).