With some difficulty, added the CSV file version of applicant list
Had to futz around with the Wayback Machine interface, and then use Chrome
developer tools to separate out the CSV file and download and save it.

Then it became clear that it used bare carriage returns (^M) to terminate the
lines, so I used

  tr "\r" "\n"

to rewrite the ^M as newlines, and saved that. That's the file that I'm
checking in here.