Parsing of root db HTML is much more flexible now
It's possible to parse both current snapshots of the IANA root db HTML
and also archived versions from the Internet Archive's Wayback Machine!

It's also possible to print out the contents of the "database" in four
different ways:

 - text
 - Lua table
 - Google sheet
 - wiki-like markup

Printing subsets of the database that either match or don't match a
sponsor is possible using the "match" and "-match" command line
keywords.